+++ database/migrations/2023_11_23_162727_modify_total_column_in_expenses_table.php
@@ -14,7 +14,7 @@ class ModifyTotalColumnInExpensesTable extends Migration
     public function up()
     {
         Schema::table('expenses', function (Blueprint $table) {
-            $table->float('amount', 8, 2)->default(0)->change();
+            $table->decimal('amount', 8, 2)->default(0)->change();
         });
     }
